+++ src/checkbox/index.ts
@@ -1,9 +1,14 @@
 import _Checkbox from './Checkbox';
+import _CheckboxGroup from './CheckboxGroup';
 
-import './style/index.js';
+import './style';
 
 export type { CheckBoxProps } from './Checkbox';
+export type { CheckboxGroupProps } from './CheckboxGroup';
+
 export * from './type';
 
 export const Checkbox = _Checkbox;
-// export default Checkbox;
+export const CheckboxGroup = _CheckboxGroup;
+
+export default Checkbox;