+++ src/image-viewer/util.ts
@@ -1,0 +1,8 @@
+export function getClientSize() {
+  const width = document.documentElement.clientWidth;
+  const height = window.innerHeight || document.documentElement.clientHeight;
+  return {
+    width,
+    height,
+  };
+}